tour/handlers: Move handlers to their own directory
The Handler* names aren't very consistent with the rest of the source
tree.
